// src/modules/chat/enums/events.enum.ts
export enum Events  {
    NEW_MESSAGE = "new-message",
    NEW_FILE = "new-file",
    ALL_ROOMS = "all-rooms",
    ALL_MESSAGES = "all-messages",
    SUCCESS = "success",
    ERROR = "error"

}

export enum SUBSCRIBED_EVENTS {
    FETCH_ROOMS = "fetch-rooms",
    MESSAGE  = "message",
    FETCH_MESSAGES = "fetch-messages",
    SEND_FILE = "send-file"
}